Refactor

+++ shlang/src/commonMain/kotlin/preprocess/ConditionEvaluationContext.kt
@@ -5,8 +5,8 @@ import tokenizer.CToken
 import types.TypeHolder
 
 
-class ConditionEvaluationContext(private val preprocessorContext: PreprocessorContext): ConstEvalContext<Int> {
-    override fun getVariable(name: CToken): Int {
+class ConditionEvaluationContext(private val preprocessorContext: PreprocessorContext): ConstEvalContext<Long> {
+    override fun getVariable(name: CToken): Long {
         val predefined = preprocessorContext.findPredefinedMacros(name.str())
         if (predefined != null) {
             return predefined.constEval()
@@ -16,7 +16,7 @@ class ConditionEvaluationContext(private val preprocessorContext: PreprocessorCo
         return if (has) 1 else 0
     }
 
-    override fun callFunction(name: CToken, args: List<Int>): Int {
+    override fun callFunction(name: CToken, args: List<Long>): Long {
         throw PreprocessorException("Unknown function '${name.str()}' in \"${name.position().filename()}\" at ${name.line()}:${name.pos()}")
     }
